--- frontend/src/components/HomepageBanner/styles.ts
@@ -1,32 +1,0 @@
-const useStyles = () => ({
-  banner: {
-    background: '#5197FF',
-    color: 'white',
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 7,
-  },
-  container: {
-    width: 'auto',
-  },
-  h2: { 
-    textAlign: 'left',
-    padding: 1,
-    margin: 0,
-    fontWeight: 200,
-    fontSize: '1.5vw',
-  },
-  mainheader: {
-    fontWeight: 700,
-    fontSize: 'max(3vw, 5vw)',
-  },
-  logo: {
-    marginRight: 20,
-    width: '20vw',
-    height: 'auto',
-  },
-});
-
-export default useStyles;
